refactor: move kv types in `xline-client` to a seperate file
