+++ xline-client/src/types/mod.rs
@@ -1,2 +1,4 @@
+/// Kv type definitions.
+pub mod kv;
 /// Watch type definitions.
 pub mod watch;
